Three narrow margins: the checksum accumulator, localtime() and chomp()
None of these is reachable today.  Each is a property that holds for a
reason outside the code that depends on it, which is the kind that stops
holding without anyone editing the line.

inet_cksum() (Coverity CID 1678600, INTEGER_OVERFLOW at the truncation).
The truncation it names is the one's complement fold of RFC 1071 and is
meant; the accumulator that feeds it is the part worth looking at.  It was
an int, and the most an IP datagram can put in it is 32768 words of 0xffff,
2147450880 -- under INT_MAX by 32767, so it never overflowed.  That margin
is a property of the IP length field, not of anything here, and the buffers
this is called on are twice what that field can describe (RECV_BUF_SIZE is
128K, src/defs.h): a caller that ever passed a buffer length instead of a
datagram length would double the worst case, and the overflow would be
signed, which is undefined rather than wraparound.  The accumulator is
uint32_t now and nleft unsigned with it, which drops the (int) cast of a
u_int parameter.  Checked rather than assumed: both versions agree on
206200 buffers, every length from 0 to 1024 and 65530 to 65535, random and
all-ones, and UBSan says "signed integer overflow: 2147450880 + 65535" for
the old one on a 128K buffer and nothing for the new.

logit() (src/debug.c).  localtime() answers NULL for a time it cannot break
down and the result went straight into thyme->tm_hour.  Losing the message
is the one thing the log path may not do, so a NULL prints the message
without the timestamp rather than dereferencing.  Coverity did not report
this one; it reports UNINIT on the va_start() four lines above, which is
the call that initialises ap and has nothing before it to read.

chomp() (src/ipc.c).  `p = str + strlen(str) - 1; while (*p == '\n') *p--
= 0;` walks off the front of the buffer for a string of nothing but
newlines, writing zeroes as it goes.  Nothing reaches it with one: every
call comes through strip(), which skips the leading run of " \t\n" first,
so the string it passes is either empty -- caught by the strlen test above
-- or starts with something else.  The same function in pimctl.c already
carries the `p >= str` test, so this is the daemon end of a fix that was
made once and not copied, and the daemon end is the one reading a socket.

No behaviour change intended anywhere: the 21 scenarios of test/lab.sh pass
on FreeBSD, and inet_cksum() is on the path of every message any of them
sends or receives.

+++ src/inet.c
@@ -164,8 +164,8 @@ uint32_t inet_parse(char *s, int n)
  */
 int inet_cksum(uint16_t *addr, u_int len)
 {
-        int sum = 0;
-        int nleft = (int)len;
+        uint32_t sum = 0;
+        u_int nleft = len;
         uint16_t *w = addr;
         uint16_t answer = 0;
 
@@ -174,6 +174,16 @@ int inet_cksum(uint16_t *addr, u_int len)
          *  we add sequential 16 bit words to it, and at the end, fold
          *  back all the carry bits from the top 16 bits into the lower
          *  16 bits.
+         *
+         *  The accumulator is unsigned because an int one is only just
+         *  wide enough: the most 0xffff words an IP datagram can hold is
+         *  65535/2, and 32768 * 0xffff is 2147450880, which clears INT_MAX
+         *  by 32767.  It is a property of the IP length field rather than
+         *  of anything here, and the buffers this is called on are twice
+         *  as large as that field can describe (RECV_BUF_SIZE, src/defs.h),
+         *  so the day something hands it a buffer length instead of a
+         *  datagram length the overflow would be signed, which is
+         *  undefined, rather than the wraparound this now has.
          */
         while (nleft > 1)  {
                 sum += *w++;
